set up seed and listen configuration
