+++ .husky/pre-commit
@@ -2,4 +2,4 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 tsc
-npm test
+npm testtsc && npm test
