+++ package.json
@@ -31,8 +31,8 @@
   "scripts": {
     "test": "jest dist/__tests__",
     "prepare": "husky install",
-    "start-dev": "nodemon node ./dist/listen.js"
-
+    "start-dev": "nodemon node ./dist/listen.js",
+    "seed-dev": "node dist/db/postgres/seed/run-seed.js"
   },
   "keywords": [],
   "author": "",
